make util internal
