+++ _util_test.ts
@@ -1,6 +1,6 @@
 // Copyright 2019-2020 Yusuke Sakurai. All rights reserved. MIT license.
 import { assertEquals } from "./vendor/https/deno.land/std/testing/asserts.ts";
-import { pathResolver } from "./util.ts";
+import { pathResolver } from "./_util.ts";
 import { group } from "./_test_util.ts";
 
 group("pathResolver", (t) => {
