+++ error.ts
@@ -10,3 +10,5 @@ export class UnexpectedEofError extends Error {
 
 /** error that is thrown when tcp connection is closed */
 export class ConnectionClosedError extends Error {}
+
+export class TimeoutError extends Error {}
