+++ mod.ts
@@ -10,7 +10,6 @@ export * from "./logger.ts";
 export * from "./matcher.ts";
 export * from "./media_types.ts";
 export * from "./middleware.ts";
-export * from "./promises.ts";
 export * from "./readers.ts";
 export * from "./responder.ts";
 export * from "./router.ts";
@@ -19,4 +18,3 @@ export * from "./serve_static.ts";
 export * from "./serveio.ts";
 export * from "./server.ts";
 export * from "./testing.ts";
-export * from "./util.ts";
